Implement the submit button from checkout form
Initialize and populate purchase object with required data to be send via the placeOrder method from checkout service.

A subscription to the placeOrder method is required to determine whether the post request succeed or not.

If successfully, then call resetShoppingCart method to rest shopping cart data, checkout form and navigate back to /products route.

Otherwise, the error is handled by logging the error message.