Update the header component to determine whether user is authenticated or not
If the user is not authenticated by okta, then display a login button in nav header.

Otherwise, if user is authenticated by okta, then retrieve and display the username along with a dropdown menu. One of the options from the dropdown menu is a sign out item to log user off when clicked.