Add account info component with a route guard to allow authenticate users
Apply the okta provided auth guard to account info route.

The oktaAuth is added to onAuthRequired flag to enable OktaAuthGuard. If a user attempts to access a protected route, then user is redirected back to okta sign in widget.

The account info component purpose is to display account related items such as personal info.