Fix issue when logging in or refreshing the page will reset the product(s) added to shopping cart
Add localStorage to the shoppingCart service to store and persist the products added to the shopping cart.

A helper method to store the current products in shopping cart via localStorage is trigger whenever the user adds a product to shopping cart.

If any component(s) inject the shopping cart service, the products stored in localStorage are retrieved, stored in local shoppingCartItems array within the service, the results are computed and published to any subscribers.

Closes #1