Add a new helper method and behvaior subject to determine whether shopping cart is valid or not
The validShoppingCart helper method determines a valid or invalid shopping cart based on the items in the cart.

If one item is invalid, the shopping cart is invalid. Otherwise, the shopping cart is invalid. The result is then emit through the behavior subject.

Introduce the validShoppingCart and shoppingCartStatus behavior subject to disallow the user from placing a order if remaining stock of atleast one product is less than quantity of product the user wishes to purchase.