Update the design and functionality of the shopping cart details component
Subscribe to shoppingCartStatus to retrieve the current status of shopping cart. Use this informtaion to disable the button navigating to checkout form and display a warning message if shopping cart is invalid.

For the design, include a message incidates to the user if product added to shopping cart has any stock remaining and update margins of certain elements.