Update some UI bugs with certain Angular components
In the product-list component, when searching for a product that doesn't exist, both the number of products found navbar and no products found warning message appears.

Instead, just show the no products found warning message. If products were found via the search, display the number of products found navbar.

In the order-history component, set a fix size for the add button.